+++ rally2-matching-system/go-example/c/sample_matching_algorithm.h
