+++ rally2-matching-system/go-example/docker/make-docker-image.sh
