+++ rally2-matching-system/go-example/docker/run-docker-image.sh
