+++ rally2-matching-system/go-example/docker/save-docker-image.sh
