+++ rally2-matching-system/go-example/main.go
@@ -10,7 +10,7 @@ import (
 	"net/http"
 	"encoding/json"
 	"unsafe"
-	"github.com/mdtf-public/rally2-matching-system/example/models"
+	"github.com/mdtf-public/rally2-matching-system/go-example/models"
 	"encoding/base64"
 	"strconv"
 	"github.com/disintegration/imaging"
