+++ rally2-matching-system/go-example/models/compare_list_request.go
