+++ rally2-matching-system/go-example/models/comparison.go
