+++ rally2-matching-system/go-example/models/info.go
